+++ api/reroutes.py
@@ -1,10 +1,12 @@
 from flask import Blueprint
-from flask_restful import Resource
+from flask_restful import Resource, Api
 
 
 reroutes = Blueprint(
     "reroutes", __name__, static_folder="../build", static_url_path="/",
 )
+
+api = Api(reroutes)
 
 
 class Index(Resource):
@@ -12,3 +14,5 @@ class Index(Resource):
 
     def get(self):
         return reroutes.send_static_file("index.html")
+
+api.add_resource(Index, "/")
