+++ requirements.txt
@@ -6,7 +6,6 @@ Flask-Cors==3.0.8
 Flask-SQLAlchemy==2.4.4
 Flask-Migrate==2.5.3
 Flask-JWT==0.3.2
-PyJWT==1.4.2
 python-dotenv==0.14.0
 pytz==2020.1
 SQLAlchemy==1.3.18
